refactored get one item
