+++ public/react/components/Item.jsx
@@ -2,6 +2,12 @@ import React from 'react';
 
 export const Item = (props) => {
 
+  const {setSelectedItem} = props
+
+  const handleItemClick = (item) => {
+		setSelectedItem(props.Item);
+	};
+
   const handleEdit = () => {
     // TODO: implement edit
     console.log("Edit button clicked");
@@ -23,8 +29,8 @@ export const Item = (props) => {
   }
 
   return <>
-  <div className="item">
-    <h3 className='name'>{props.Item.name}</h3>
+  <div className="item" >
+    <h3 className='name' onClick={handleItemClick}>{props.Item.name}</h3>
     <h4 className='category'>{props.Item.category}</h4>
     <img src={props.Item.image} alt={props.Item.name} />
     <p className='description'>{props.Item.description}</p>
